+++ Editor/Editors/FloaterEditor.cs
@@ -5,28 +5,59 @@ using UnityEngine;
 namespace GOcean
 {
     [CustomEditor(typeof(Floater))]
+    [CanEditMultipleObjects]
     public class FloaterEditor : Editor
     {
-        private SerializedProperty
-            uniformSamplerIterations,
-            uniformForce,
-            uniformRadius,
-            samplerIterations,
-            force,
-            radius;
+        private const string EDITOR_PREFS_STRING_KEY = "GOCEAN_FLOATER_EDITOR_";
 
         private SerializedProperty buoyancyInfluences;
         private ReorderableList buoyancyInfluencesList;
         private Floater floater;
-
-        private bool prevToolHiddenState = false;
+        private bool prevToolHiddenState;
+        private uint iterations;
+        private float force;
+        private float radius;
 
         private void OnEnable()
         {
             floater = target as Floater;
 
             GetPropertyReferences();
+            LoadData();
+            SetupBuoyancyInfluencesList();
 
+            prevToolHiddenState = Tools.hidden;
+        }
+
+        private void OnDisable()
+        {
+            SaveData();
+            CleanupBuoyancyInfluencesList();
+
+            Tools.hidden = prevToolHiddenState;
+        }
+
+        private void GetPropertyReferences()
+        {
+            buoyancyInfluences = serializedObject.FindProperty("buoyancyInfluences");
+        }
+
+        private void LoadData()
+        {
+            iterations = (uint)EditorPrefs.GetInt($"{EDITOR_PREFS_STRING_KEY}iterations", 0);
+            force = EditorPrefs.GetFloat($"{EDITOR_PREFS_STRING_KEY}force", 1f);
+            radius = EditorPrefs.GetFloat($"{EDITOR_PREFS_STRING_KEY}radius", 1f);
+        }
+
+        private void SaveData()
+        {
+            EditorPrefs.SetInt($"{EDITOR_PREFS_STRING_KEY}iterations", (int)iterations);
+            EditorPrefs.SetFloat($"{EDITOR_PREFS_STRING_KEY}force", force);
+            EditorPrefs.SetFloat($"{EDITOR_PREFS_STRING_KEY}radius", radius);
+        }
+
+        private void SetupBuoyancyInfluencesList()
+        {
             if (buoyancyInfluencesList == null)
             {
                 buoyancyInfluencesList = new ReorderableList(serializedObject, buoyancyInfluences);
@@ -35,17 +66,13 @@ namespace GOcean
             buoyancyInfluencesList.drawElementCallback += DrawBuoyancyInfluencesListElement;
             buoyancyInfluencesList.elementHeightCallback += GetDrawBuoyancyInfluencesListElementHeight;
             buoyancyInfluencesList.drawHeaderCallback += DrawBuoyancyInfluencesListHeader;
-
-            prevToolHiddenState = Tools.hidden;
         }
 
-        private void OnDisable()
+        private void CleanupBuoyancyInfluencesList()
         {
             buoyancyInfluencesList.drawElementCallback -= DrawBuoyancyInfluencesListElement;
             buoyancyInfluencesList.elementHeightCallback -= GetDrawBuoyancyInfluencesListElementHeight;
             buoyancyInfluencesList.drawHeaderCallback -= DrawBuoyancyInfluencesListHeader;
-
-            Tools.hidden = prevToolHiddenState;
         }
 
         public override void OnInspectorGUI()
@@ -54,27 +81,45 @@ namespace GOcean
 
             using (new GUILayout.HorizontalScope())
             {
-                EditorGUILayout.PropertyField(uniformSamplerIterations);
-                GUI.enabled = uniformSamplerIterations.boolValue;
-                EditorGUILayout.PropertyField(samplerIterations, GUIContent.none);
-                GUI.enabled = true;
+                bool update = GUILayout.Button("Set Iterations", GUILayout.Width(200f));
+                iterations = (uint)EditorGUILayout.IntField((int)iterations);
+                if (update)
+                {
+                    for (int i = 0; i < buoyancyInfluencesList.count; i++)
+                    {
+                        buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(i).FindPropertyRelative("oceanSampler").FindPropertyRelative("iterations").uintValue = iterations;
+                    }
+                }
             }
             using (new GUILayout.HorizontalScope())
             {
-                EditorGUILayout.PropertyField(uniformForce);
-                GUI.enabled = uniformForce.boolValue;
-                EditorGUILayout.PropertyField(force, GUIContent.none);
-                GUI.enabled = true;
+                bool update = GUILayout.Button("Set Force", GUILayout.Width(200f));
+                force = EditorGUILayout.FloatField(force);
+                if (update)
+                {
+                    for (int i = 0; i < buoyancyInfluencesList.count; i++)
+                    {
+                        buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(i).FindPropertyRelative("force").floatValue = force;
+                    }
+                }
             }
             using (new GUILayout.HorizontalScope())
             {
-                EditorGUILayout.PropertyField(uniformRadius);
-                GUI.enabled = uniformRadius.boolValue;
-                EditorGUILayout.PropertyField(radius, GUIContent.none);
-                GUI.enabled = true;
+                bool update = GUILayout.Button("Set Radius", GUILayout.Width(200f));
+                radius = EditorGUILayout.FloatField(radius);
+                if (update)
+                {
+                    for (int i = 0; i < buoyancyInfluencesList.count; i++)
+                    {
+                        buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(i).FindPropertyRelative("radius").floatValue = radius;
+                    }
+                }
             }
 
-            buoyancyInfluencesList.DoLayoutList();
+            if (targets.Length < 2)
+            {
+                buoyancyInfluencesList.DoLayoutList();
+            }
 
             serializedObject.ApplyModifiedProperties();
         }
@@ -85,6 +130,8 @@ namespace GOcean
             bool mouseDrag = Event.current.type == EventType.MouseDrag;
             bool mouseButtonLeft = Event.current.button == 0;
             bool buoyancyObjectIsSelected = buoyancyInfluencesList.selectedIndices.Count > 0;
+
+            //SerializedObject[] targetObjects = targets.Select(t => new SerializedObject(t)).ToArray();
 
             Color prevColor = Handles.color;
 
@@ -121,13 +168,13 @@ namespace GOcean
                     }
                 }
 
-                SerializedProperty currentElement = buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(i);
-                Vector3 localPosition = currentElement.FindPropertyRelative("localPosition").vector3Value;
-                Vector3 worldPosition = floater.transform.localToWorldMatrix.MultiplyPoint(localPosition);
-                float radius = uniformRadius.boolValue ? this.radius.floatValue : currentElement.FindPropertyRelative("radius").floatValue;
+                BuoyancyInfluence bInfluence = floater.buoyancyInfluences[i];
+                Vector3 worldPosition = floater.transform.localToWorldMatrix.MultiplyPoint(bInfluence.GetLocalPosition());
+                float radius = bInfluence.GetRadius();
+                float diameter = radius * 2f;
 
                 // select new buoyancy object if it was clicked on
-                if (Handles.Button(worldPosition, Quaternion.identity, radius, radius * 0.5f, Handles.SphereHandleCap))
+                if (Handles.Button(worldPosition, Quaternion.identity, diameter, radius, Handles.SphereHandleCap))
                 {
                     buoyancyInfluencesList.Select(i);
                     this.Repaint();
@@ -139,68 +186,47 @@ namespace GOcean
             //      draw scale / move handle
             if (buoyancyObjectIsSelected)
             {
-                SerializedProperty selectedElement = buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(buoyancyInfluencesList.selectedIndices[0]);
-                SerializedProperty positionElement = selectedElement.FindPropertyRelative("localPosition");
-                SerializedProperty radiusElement = selectedElement.FindPropertyRelative("radius");
-
-                Vector3 worldPosition = floater.transform.localToWorldMatrix.MultiplyPoint(positionElement.vector3Value);
-
-                serializedObject.Update();
+                BuoyancyInfluence bInfluence = floater.buoyancyInfluences[buoyancyInfluencesList.selectedIndices[0]];
+                Vector3 worldPosition = floater.transform.localToWorldMatrix.MultiplyPoint(bInfluence.GetLocalPosition());
+                float radius = bInfluence.GetRadius();
+                float diameter = radius * 2f;
 
                 switch (Tools.current)
                 {
                     case Tool.Move:
                         Vector3 newPosition = Handles.PositionHandle(worldPosition, floater.transform.rotation);
-                        if (mouseDrag && mouseButtonLeft)
+                        if (newPosition != worldPosition)
                         {
-                            positionElement.vector3Value = floater.transform.worldToLocalMatrix.MultiplyPoint(newPosition);
+                            Undo.RecordObject(target, "Move Buoyancy Object");
+                            Vector3 localPosition = floater.transform.worldToLocalMatrix.MultiplyPoint(newPosition);
+                            bInfluence.SetLocalPosition(localPosition);
+                            bInfluence.UpdateWorldPosition(floater.transform.localToWorldMatrix);
                         }
                         break;
                     case Tool.Scale:
-                        if (uniformRadius.boolValue)
+                        Vector3 scale = Handles.ScaleHandle(new Vector3(radius, radius, radius), worldPosition, floater.transform.rotation);
+                        for (int i = 0; i < 3; i++)
                         {
-                            break;
-                        }
-                        Vector3 scale = Handles.ScaleHandle(new Vector3(radiusElement.floatValue, radiusElement.floatValue, radiusElement.floatValue), worldPosition, floater.transform.rotation);
-                        if (mouseDrag && mouseButtonLeft)
-                        {
-                            for (int i = 0; i < 3; i++)
+                            if (scale[i] != radius)
                             {
-                                if (scale[i] != radiusElement.floatValue)
-                                {
-                                    scale.x = scale[i];
-                                    break;
-                                }
+                                Undo.RecordObject(target, "Scale Buoyancy Object");
+                                bInfluence.SetRadius(Mathf.Max(scale[i], 0f));
+                                break;
                             }
-                            radiusElement.floatValue = Mathf.Max(scale.x, 0.001f);
                         }
                         break;
                     default:
                         break;
                 }
 
-                serializedObject.ApplyModifiedProperties();
-
                 if (Event.current.type == EventType.Repaint)
                 {
                     Handles.color = new Color(0.7f, 0.2f, 0.1f, 0.5f);
-                    float radius = uniformRadius.boolValue ? this.radius.floatValue : radiusElement.floatValue;
-                    Handles.SphereHandleCap(-1, worldPosition, Quaternion.identity, radius, EventType.Repaint);
+                    Handles.SphereHandleCap(-1, worldPosition, Quaternion.identity, diameter, EventType.Repaint);
                 }
             }
 
             Handles.color = prevColor;
-        }
-
-        private void GetPropertyReferences()
-        {
-            uniformSamplerIterations = serializedObject.FindProperty("uniformSamplerIterations");
-            uniformForce = serializedObject.FindProperty("uniformForce");
-            uniformRadius = serializedObject.FindProperty("uniformRadius");
-            samplerIterations = serializedObject.FindProperty("samplerIterations");
-            force = serializedObject.FindProperty("force");
-            radius = serializedObject.FindProperty("radius");
-            buoyancyInfluences = serializedObject.FindProperty("buoyancyInfluences");
         }
 
         private void DrawBuoyancyInfluencesListHeader(Rect rect)
@@ -217,9 +243,9 @@ namespace GOcean
         private void DrawBuoyancyInfluencesListElement(Rect rect, int index, bool isActive, bool isFocused)
         {
             SerializedProperty property = buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(index);
-            SerializedProperty currentProp;
+            EditorGUI.PropertyField(rect, property);
 
-            float h = 0f;
+            /*float h = 0f;
             Rect r = new Rect(rect.x, rect.y, rect.width, h);
 
             GUI.enabled = !uniformSamplerIterations.boolValue;
@@ -228,14 +254,14 @@ namespace GOcean
             h = EditorGUI.GetPropertyHeight(currentProp);
             r.height = h;
             EditorGUI.PropertyField(r, currentProp);
-            
+
             GUI.enabled = !uniformForce.boolValue;
             currentProp = property.FindPropertyRelative("force");
             r.y += h + BuoyancyInfluencePropertyDrawer.PADDING_HEIGHT;
             h = EditorGUI.GetPropertyHeight(currentProp);
             r.height = h;
             EditorGUI.PropertyField(r, currentProp);
-            
+
             GUI.enabled = !uniformRadius.boolValue;
             currentProp = property.FindPropertyRelative("radius");
             r.y += h + BuoyancyInfluencePropertyDrawer.PADDING_HEIGHT;
@@ -249,7 +275,15 @@ namespace GOcean
             r.y += h + BuoyancyInfluencePropertyDrawer.PADDING_HEIGHT;
             h = EditorGUI.GetPropertyHeight(currentProp);
             r.height = h;
-            EditorGUI.PropertyField(r, currentProp);
+            EditorGUI.PropertyField(r, currentProp);*/
+        }
+
+        private void UpdateBuoyancyInfluencesListProperty(string propertyName)
+        {
+            for (int i = 0; i < buoyancyInfluencesList.count; i++)
+            {
+                SerializedProperty prop = buoyancyInfluencesList.serializedProperty.GetArrayElementAtIndex(i).FindPropertyRelative(propertyName);
+            }
         }
     }
 }