+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.1.0
+- Proper physics
+- Floater component that interfaces with rigidbody
+
 ## 1.0.1 (2025-02-15)
 - Terrain system bug fixes
 - Debug warnings for mismatched terrain data